Refactor: Replace AI modules with perfect versions
This commit replaces the original AI content generation, monitoring, and provider modules with their "perfect" counterparts. This likely involves refactoring and renaming of files and imports to reflect the new structure.
